ml-services/requirements.txt
fastapi
uvicorn
ultralytics
torch
transformers
Pillow
requests
python-multipart
sentence-transformers
scikit-learn
cloudinary 
python-dotenv 
opencv-python 
numpy 
Pillow
pandas
joblib
xgboost
overpy